Added names and numbers top of code and started a bit of backpack

+++ Game/src/game/Main.java
@@ -1,5 +1,6 @@
 package game;
-
+// Simon Ung 101032525
+// Chi Calvin Nguyen 101203877
 import java.util.Scanner;
 
 public class Main {
@@ -61,7 +62,7 @@ public class Main {
             if (si != null) {
 
                 p.buy(si.item);
-                p.withdraw(si.item.cost);
+                p.withdraw((int) si.item.cost);
                 si.numberInStock--;
 
             } else {
